CLEANUP fully functional server with modular routing CLEANUP adds comments and link to the wiki page for Map.react
CLEANUP refactors client/server api endpoints into their own routing

CLEAN UP refactors server for dataProcessing functions used when importing and processing new data

CLEANUP comments out all files that are about to be deleted

CLEAN UP deletes server files that are no longer being used

CLEAN UP moves sql schema to server/db and sql queries to server/dataProcessing
