+++ client/app/components/Map.react.js
@@ -1,6 +1,8 @@
 var React = require('react');
-var languageColors = require('./languageColors');
+//languageColors is the object that maps from a language string to a color hex.
+var languageColors = require('./languageColors'); 
 
+//sample data so we can see the format 
 // var mockData = {
 //   "BRN":{"fillKey":"JavaScript","allLangs":[["JavaScript",3],["Visual Basic",3],["PHP",1]]},
 //   "THA":{"fillKey":"JavaScript","allLangs":[["JavaScript",2574],["Python",1108],["PHP",922],["Java",862],["CSS",717],["Ruby",686],["Objective-C",583],["C",356],["Shell",306],["Go",295],["C++",253],["Swift",146],["CoffeeScript",138],["C#",115],["Logos",95],["VimL",85],["Scala",84],["Perl",51],["Clojure",45],["Emacs Lisp",43],["Makefile",36],["Lua",29],["Bison",28],["Puppet",27],["R",26],["TeX",25],["Arduino",18],["Elixir",18],["Common Lisp",16],["Objective-C++",14],["PowerShell",13],["Erlang",11],["Haskell",9],["Haxe",9],["Rust",9],["IDL",8],["XSLT",8],["TypeScript",6],["Frege",5],["Groovy",5],["Vala",5],["Hy",4],["VCL",4],["Assembly",2],["F#",2],["Matlab",2],["Visual Basic",2],["Elm",1],["Nimrod",1],["Racket",1]]},
@@ -9,15 +11,19 @@ var languageColors = require('./languageColors');
 
 var Map = React.createClass({
 
+  //Please see wiki page for more detailed information:
+  //https://github.com/CulturedCheese/thesis-project/wiki/DataMaps
+
   render: function() {
     var map = new Datamap({
       element: document.getElementById('d3Map'),
-      fills: languageColors,
-      data: this.props.countryData,
+      fills: languageColors, //mapping file from language to the color code. it's a long file so we're saving it elsewhere. 
+      data: this.props.countryData,  //this is the data that is attached to each country
       geographyConfig: {
         popupTemplate: function(geo, data) {
           //TODO: format this. right now it's just putting the whole 'data.allLangs' object into the hover
           //clearly we want to have this formatted more intelligently
+          //we should be able to make this a separate React component
           return ['<div class="hoverinfo"><strong>',
                   'Number of things in ' + geo.properties.name,
                   ': ' + data.allLangs,
