+++ server/api/1/api-router.js
@@ -1,0 +1,14 @@
+var express = require('express');
+var apiRouter = express.Router();
+var apiLogic = require('./apiLogic.js');
+
+apiRouter.get('/allCountriesAllLanguages', function(req,res) {
+  apiLogic.allCountriesAllLanguages(req,res);
+});
+
+apiRouter.get('/countriesForLanguage', function(req,res) {
+  console.log('heard a request to countriesForLanguage');
+  apiLogic.countriesForLanguage(req,res);
+});
+
+module.exports = apiRouter;
