+++ server/dataProcessing/sqlQueries/LoadData.sql
