+++ server/dataProcessing/sqlQueries/aggCountriesToTop10.sql
