+++ server/dataProcessing/sqlQueries/groupLocations.sql
