+++ server/dataProcessing/sqlQueries/moveDataToPlaces.sql
