--- server/db/githubToMysqlMap.js
@@ -1,218 +1,0 @@
-//MySQL is pretyt picky in terms of what you can put into a column name
-//and github is wildly unopinionated in term sof what you can put into a language name
-//this gets the two to agree on items such as #, spaces, -, and even the apostrophe in Cap'n something or other
-//to use, you need to create a variable, such as langName, and set it equal to a string version of the github language name
-//then look up that variable in this object: langMap[langName], and that will give you what the proper column is in MySQL
-module.exports= {
-  "C#": 'Csharp',
-  "Prolog": 'Prolog',
-  "C": 'C',
-  "Python": 'Python',
-  "Objective-C": 'Objective_C',
-  "CSS": 'CSS',
-  "JavaScript": 'JavaScript',
-  "TeX": 'TeX',
-  "PHP": 'PHP',
-  "Java": 'Java',
-  "FORTRAN": 'FORTRAN',
-  "Go": 'Go',
-  "C++": 'Cplusplus',
-  "Swift": 'Swift',
-  "Ruby": 'Ruby',
-  "VimL": 'VimL',
-  "Lua": 'Lua',
-  "Emacs Lisp": 'Emacs_Lisp',
-  "OCaml": 'OCaml',
-  "R": 'R',
-  "Propeller Spin": 'Propeller_Spin',
-  "Shell": 'Shell',
-  "Arduino": 'Arduino',
-  "Pascal": 'Pascal',
-  "DM": 'DM',
-  "Scala": 'Scala',
-  "ASP": 'ASP',
-  "Visual Basic": 'Visual_Basic',
-  "Assembly": 'Assembly',
-  "Perl": 'Perl',
-  "CoffeeScript": 'CoffeeScript',
-  "Clojure": 'Clojure',
-  "Makefile": 'Makefile',
-  "Rust": 'Rust',
-  "Racket": 'Racket',
-  "Haskell": 'Haskell',
-  "Papyrus": 'Papyrus',
-  "Common Lisp": 'Common_Lisp',
-  "Puppet": 'Puppet',
-  "TypeScript": 'TypeScript',
-  "D": 'D',
-  "Smalltalk": 'Smalltalk',
-  "Matlab": 'Matlab',
-  "Elixir": 'Elixir',
-  "Groovy": 'Groovy',
-  "Hack": 'Hack',
-  "Frege": 'Frege',
-  "Erlang": 'Erlang',
-  "Julia": 'Julia',
-  "LiveScript": 'LiveScript',
-  "Mercury": 'Mercury',
-  "Elm": 'Elm',
-  "Scheme": 'Scheme',
-  "Objective-C++": 'Objective_Cplusplus',
-  "XSLT": 'XSLT',
-  "SQF": 'SQF',
-  "Tcl": 'Tcl',
-  "Processing": 'Processing',
-  "Perl6": 'Perl6',
-  "PowerShell": 'PowerShell',
-  "PureScript": 'PureScript',
-  "Gnuplot": 'Gnuplot',
-  "Vala": 'Vala',
-  "LoomScript": 'LoomScript',
-  "Nimrod": 'Nimrod',
-  "Verilog": 'Verilog',
-  "ActionScript": 'ActionScript',
-  "SuperCollider": 'SuperCollider',
-  "Haxe": 'Haxe',
-  "PigLatin": 'PigLatin',
-  "Idris": 'Idris',
-  "VHDL": 'VHDL',
-  "SystemVerilog": 'SystemVerilog',
-  "Dart": 'Dart',
-  "Parrot": 'Parrot',
-  "Cuda": 'Cuda',
-  "Bison": 'Bison',
-  "Mathematica": 'Mathematica',
-  "xBase": 'xBase',
-  "Inform 7": 'Inform_7',
-  "Nix": 'Nix',
-  "Coq": 'Coq',
-  "AGS Script": 'AGS_Script',
-  "SourcePawn": 'SourcePawn',
-  "OpenSCAD": 'OpenSCAD',
-  "ColdFusion": 'ColdFusion',
-  "ANTLR": 'ANTLR',
-  "SQL": 'SequelDB',
-  "Xtend": 'Xtend',
-  "AutoIt": 'AutoIt',
-  "OpenEdge ABL": 'OpenEdge_ABL',
-  "ATS": 'ATS',
-  "F#": 'Fsharp',
-  "Max": 'Max',
-  "AppleScript": 'AppleScript',
-  "Eiffel": 'Eiffel',
-  "Gosu": 'Gosu',
-  "Game Maker Language": 'Game_Maker_Language',
-  "Standard ML": 'Standard_ML',
-  "Pure Data": 'Pure_Data',
-  "AutoHotkey": 'AutoHotkey',
-  "Ada": 'Ada',
-  "SAS": 'SAS',
-  "Chapel": 'Chapel',
-  "Forth": 'Forth',
-  "UnrealScript": 'UnrealScript',
-  "Stata": 'Stata',
-  "Logos": 'Logos',
-  "M": 'M',
-  "Arc": 'Arc',
-  "IDL": 'IDL',
-  "JSONiq": 'JSONiq',
-  "BitBake": 'BitBake',
-  "Objective-J": 'Objective_J',
-  "BlitzBasic": 'BlitzBasic',
-  "XQuery": 'XQuery',
-  "Squirrel": 'Squirrel',
-  "Apex": 'Apex',
-  "Nemerle": 'Nemerle',
-  "J": 'J',
-  "Kotlin": 'Kotlin',
-  "nesC": 'nesC',
-  "ooc": 'ooc',
-  "Agda": 'Agda',
-  "Factor": 'Factor',
-  "Crystal": 'Crystal',
-  "GAP": 'GAP',
-  "VCL": 'VCL',
-  "Rebol": 'Rebol',
-  "LabVIEW": 'LabVIEW',
-  "Ceylon": 'Ceylon',
-  "CLIPS": 'CLIPS',
-  "MoonScript": 'MoonScript',
-  "ABAP": 'ABAP',
-  "GAMS": 'GAMS',
-  "Component Pascal": 'Component_Pascal',
-  "COBOL": 'COBOL',
-  "AspectJ": 'AspectJ',
-  "Red": 'Red',
-  "Pan": 'Pan',
-  "Bluespec": 'Bluespec',
-  "Nu": 'Nu',
-  "Isabelle": 'Isabelle',
-  "EmberScript": 'EmberScript',
-  "Shen": 'Shen',
-  "Awk": 'Awk',
-  "Augeas": 'Augeas',
-  "Xojo": 'Xojo',
-  "Io": 'Io',
-  "LookML": 'LookML',
-  "Grammatical Framework": 'Grammatical_Framework',
-  "PAWN": 'PAWN',
-  "DOT": 'DOT',
-  "Boo": 'Boo',
-  "Bro": 'Bro',
-  "LSL": 'LSL',
-  "Lasso": 'Lasso',
-  "Monkey": 'Monkey',
-  "NetLogo": 'NetLogo',
-  "Scilab": 'Scilab',
-  "Delphi": 'Delphi',
-  "Brightscript": 'Brightscript',
-  "KRL": 'KRL',
-  "Golo": 'Golo',
-  "Grace": 'Grace',
-  "FLUX": 'FLUX',
-  "APL": 'APL',
-  "Glyph": 'Glyph',
-  "GDScript": 'GDScript',
-  "Fantom": 'Fantom',
-  "Logtalk": 'Logtalk',
-  "Alloy": 'Alloy',
-  "PogoScript": 'PogoScript',
-  "XML": 'XML',
-  "Oz": 'Oz',
-  "IGOR Pro": 'IGOR_Pro',
-  "wisp": 'wisp',
-  "Hy": 'Hy',
-  "Slash": 'Slash',
-  "Zephir": 'Zephir',
-  "Pike": 'Pike',
-  "Opal": 'Opal',
-  "REALbasic": 'REALbasic',
-  "BlitzMax": 'BlitzMax',
-  "XC": 'XC',
-  "Rouge": 'Rouge',
-  "Fancy": 'Fancy',
-  "Ox": 'Ox',
-  "Clean": 'Clean',
-  "RobotFramework": 'RobotFramework',
-  "Dylan": 'Dylan',
-  "PureBasic": 'PureBasic',
-  "TXL": 'TXL',
-  "Jasmin": 'Jasmin',
-  "Mirah": 'Mirah',
-  "XProc": 'XProc',
-  "Thrift": 'Thrift',
-  "Ragel in Ruby Host": 'Ragel_in_Ruby_Host',
-  "Turing": 'Turing',
-  "Ecl": 'Ecl',
-  "DCPU-16 ASM": 'DCPU_16_ASM',
-  "Cycript": 'Cycript',
-  "Self": 'Self',
-  "Cool": 'Cool',
-  "Oxygene": 'Oxygene',
-  "Volt": 'Volt',
-  "Opa": 'Opa',
-  "E": 'E',
-  "WebIDL": 'WebIDL',
-  "Cap'n Proto": 'Capn_Proto',
-  "eC": 'eC'
-};
