+++ server/db/sqlSchema/countriesAggregatedSchema.sql
