+++ server/db/sqlSchema/placesSchema.sql
