+++ server/db/sqlSchema/placesWithGeoSchema.sql
