+++ server/db/sqlSchema/rawDataSchema.sql
