+++ server/db/sqlSchema/usersAggregatedSchema.sql
