--- server/index.html
@@ -1,7 +1,0 @@
-<script src="//cdnjs.cloudflare.com/ajax/libs/d3/3.5.3/d3.min.js"></script>
-<script src="//cdnjs.cloudflare.com/ajax/libs/topojson/1.6.9/topojson.min.js"></script>
-<script src="/datamaps.world.min.js"></script>
-<div id="container" style="position: relative; width: 500px; height: 300px;"></div>
-
-
-<script src="/mapScripts.js"></script>
