--- server/mapScripts.js
@@ -1,1 +1,0 @@
-var map = new Datamap({element: document.getElementById('container')});
